Updated desing in bill overview
