+++ app/src/main/java/asay/asaymobile/fragments/BillCommentsFragment.java
@@ -1,25 +1,33 @@
 package asay.asaymobile.fragments;
 
 import android.animation.ObjectAnimator;
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v7.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.ImageButton;
 import android.widget.TextView;
 
 import java.util.ArrayList;
 
 import asay.asaymobile.ForumContract;
 import asay.asaymobile.R;
+import asay.asaymobile.UserContract;
+import asay.asaymobile.activities.VoteActivity;
 import asay.asaymobile.model.ArgumentType;
 import asay.asaymobile.model.BillDTO;
 import asay.asaymobile.model.CommentDTO;
 import asay.asaymobile.model.UserDTO;
 import asay.asaymobile.presenter.ForumPresenter;
+import asay.asaymobile.presenter.UserPresenter;
+import butterknife.ButterKnife;
 
-public class BillCommentsFragment extends Fragment implements View.OnClickListener, ForumContract.View{
+public class BillCommentsFragment extends Fragment implements View.OnClickListener, ForumContract.View, UserContract.View{
 
     TextView BillDesc;
     String BillDescOrg;
@@ -36,6 +44,11 @@ public class BillCommentsFragment extends Fragment implements View.OnClickListen
     private BillDTO bill;
     private double userId = 1;
     ForumPresenter presenter;
+    ImageButton sub;
+    Button vote;
+    boolean isSub = false;
+    private UserDTO user;
+    private UserPresenter uPresenter;
 
     public BillCommentsFragment() {
         // Required empty public constructor
@@ -52,8 +65,16 @@ public class BillCommentsFragment extends Fragment implements View.OnClickListen
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        bill = getArguments().getParcelable("bill");
+        Bundle bundle = new Bundle();
+        bundle.putParcelable("bill", bill);
+        setArguments(bundle);
+
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_bill_comments, container, false);
+
+
     }
 
     @Override
@@ -79,6 +100,21 @@ public class BillCommentsFragment extends Fragment implements View.OnClickListen
         expArg2 = (TextView) rootView.findViewById(R.id.expandArgAgainst);
         expArg2.setOnClickListener(this);
         expArg2.setVisibility(View.INVISIBLE);
+
+        sub = (ImageButton) rootView.findViewById(R.id.subbtn);
+        sub.setOnClickListener(this);
+
+        uPresenter = new UserPresenter(this);
+        uPresenter.getUser(userId);
+        String billTitle = bill.getNumber().concat(": ").concat(bill.getTitleShort());
+        TextView header = (TextView) rootView.findViewById(R.id.headerBill);
+        header.setText(billTitle);
+        vote = (Button) rootView.findViewById(R.id.buttonVote);
+        vote.setOnClickListener(this);
+
+        if (((AppCompatActivity) getActivity()).getSupportActionBar() != null)
+            ((AppCompatActivity) getActivity()).getSupportActionBar().setTitle(billTitle);
+
 
         super.onViewCreated(rootView, savedInstanceState);
 
@@ -156,6 +192,16 @@ public class BillCommentsFragment extends Fragment implements View.OnClickListen
                     arg1.setOnClickListener(null);
                 }
                 break;
+            case R.id.buttonVote:
+                Intent voteIntent = new Intent(this.getActivity(), VoteActivity.class);
+                voteIntent.putExtra("bill", bill);
+                startActivity(voteIntent);
+                break;
+
+            case R.id.subbtn :
+                toogleFavorite(sub);
+                break;
+
         }
     }
 
@@ -251,10 +297,46 @@ public class BillCommentsFragment extends Fragment implements View.OnClickListen
             if (expArg2 != null)
                 expArg2.setVisibility(View.INVISIBLE);
         }
+
+
+    }
+
+    private void toogleFavorite(ImageButton button){
+        ArrayList<Integer> billSaved = user.getbillsSaved();
+
+        if(isSub){
+            for(int i = billSaved.size() -1 ; i >= 0; i-- ){
+                if(billSaved.get(i).equals(bill.getId()))
+                    billSaved.remove(i);
+            }
+            sub.setImageResource(R.drawable.ic_star);
+        }else {
+            billSaved.add(bill.getId());
+            sub.setImageResource(R.drawable.ic_star_border);
+        }
+        uPresenter.UpdateFavorites(userId,billSaved);
+    }
+
+    @Override
+    public void refreshUser(UserDTO user) {
+        System.out.println("number of bills saved: " + user.getbillsSaved().size());
+        isSub = user.getbillsSaved().contains(bill.getId());
+        try{
+            if (isSub)
+                sub.setImageResource(R.drawable.ic_star);
+            else{
+                sub.setImageResource(R.drawable.ic_star_border);
+            }
+        } catch (Exception e){
+            System.out.println(e.getMessage());
+        }
+        this.user = user;
     }
 
     @Override
     public void refreshUsers(ArrayList<UserDTO> users) {
 
     }
+
+
 }
